converts gabor filter to new class and fixes launch file.
