+++ line_detection/launch/gabor.launch
@@ -1,6 +1,6 @@
 <launch>
 
-  <node pkg="line_detection" name="gabor_linedetection" type="gabor-linedetection.py" output="screen">
+  <node pkg="line_detection" name="gabor" type="gabor.py" output="screen">
     
     <param name="use_mono" type="bool" value="false" />
     <param name="use_compressed_format" type="bool" value="true" />
