+++ line_detection/src/gabor.py
@@ -1,0 +1,67 @@
+#!/usr/bin/env python
+import sys
+import cv2
+import rospy
+from dynamic_reconfigure.server import Server
+from lane_detection import LaneDetection
+from line_detection.cfg import LineDetectionConfig
+
+###############################################################################
+# Chicago Engineering Design Team
+# Gabor filter using Python OpenCV for autonomous robot Scipio
+# (IGVC competition).
+# @author Basheer Subei
+# @email basheersubei@gmail.com
+
+
+class Gabor(LaneDetection):
+    image_height = 0
+    image_width = 0
+    roi_top_left_x = 0
+    roi_top_left_y = 0
+    roi_width = 2000
+    roi_height = 2000
+
+    def __init__(self, namespace, node_name):
+        LaneDetection.__init__(self, namespace, node_name)
+
+    # this is what gets called when an image is recieved
+    def image_callback(self, ros_image):
+
+        cv2_image = LaneDetection.ros_to_cv2_image(self, ros_image)
+        # apply Gabor filter
+        gabor_kernel = cv2.getGaborKernel((self.gabor_ksize, self.gabor_ksize),
+                                          self.gabor_sigma,
+                                          self.gabor_theta * math.pi / 180,
+                                          self.gabor_lambda,
+                                          self.gabor_gamma)
+
+        final_image = cv2.filter2D(cv2_image, -1, gabor_kernel)
+
+        final_image_message = LaneDetection.cv2_to_ros_message(
+            self, final_image
+        )
+        # publishes final image message in ROS format
+        self.line_image_pub.publish(final_image_message)
+    # end image_callback()
+
+
+def main(args):
+    node_name = "gabor"
+    namespace = rospy.get_namespace()
+    if namespace == "/":
+        namespace = ""
+
+    # create a gabor object
+    g = Gabor(namespace, node_name)
+
+    # start the line_detector node and start listening
+    rospy.init_node("gabor", anonymous=True)
+
+    # starts dynamic_reconfigure server
+    srv = Server(LineDetectionConfig, g.reconfigure_callback)
+    rospy.spin()
+    cv2.destroyAllWindows()
+
+if __name__ == '__main__':
+    main(sys.argv)
